fix: use f32 precision across all matmul and RoPE paths
Combined precision fixes to match GPU behavior:
- Blockwise quantized matmul: f32 cross-block accumulation
- LinearBase: f32 preferred_element_type with bf16 output cast
- Logits: cast to f32 after dot product
- RoPE: keep inv_freq, cos, sin in f32; cast output back

+++ python/sgl_jax/srt/kernels/quantized_matmul/quantized_matmul_kernels/blockwise_kernel.py
@@ -103,7 +103,7 @@ def quantized_matmul_kernel(
     save_x_q = quantize_activation and n_in == 1 and n_out > 1
 
     # TODO(amandaliang): Make this configurable.
-    acc_dtype = jnp.bfloat16
+    acc_dtype = jnp.float32
     if quantize_activation and jnp.issubdtype(w_q.dtype, jnp.integer):
         acc_dtype = jnp.int32
